move admin functions out of races/ into admin/ (refactor)
