+++ buggy_race_server/admin/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 """The admin module."""
-from . import views
+from . import views, views_races
