+++ buggy_race_server/app.py
@@ -156,6 +156,7 @@ def register_extensions(app):
 def register_blueprints(app):
     """Register Flask blueprints."""
     app.register_blueprint(admin.views.blueprint)
+    app.register_blueprint(admin.views_races.blueprint)
     app.register_blueprint(public.views.blueprint)
     app.register_blueprint(user.views.blueprint)
     app.register_blueprint(buggy.views.blueprint)
