+++ buggy_race_server/templates/admin/_nav.html
@@ -4,7 +4,7 @@
       <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('admin.admin')}}">Dashboard</a>
       <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('admin.list_users')}}">Users</a>
       <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('admin.list_buggies')}}">Buggies</a>
-      <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('race.list_races')}}">Races</a>
+      <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('admin_race.list_races')}}">Races</a>
       {% if current_user.is_administrator %}
         <a class="btn btn-outline-secondary btn-admin btn-jump" href="{{ url_for('admin.list_announcements')}}">Announce</a>
       {% endif %}
