+++ buggy_race_server/templates/admin/races.html
@@ -39,11 +39,11 @@
           <tr class="race-visible-{{ race.is_visible }}">
             {% if current_user.is_administrator %}
               <td class="btn-collection">
-                <a class="btn btn-outline-secondary btn-admin btn-sm btn-jump" href="{{ url_for('race.edit_race', race_id=race.id) }}">Edit</a>
+                <a class="btn btn-outline-secondary btn-admin btn-sm btn-jump" href="{{ url_for('admin_race.edit_race', race_id=race.id) }}">Edit</a>
                 {% if race.result_log_url %}
                   <a class="btn btn-outline-secondary btn-admin btn-sm btn-jump" href="{{ url_for('race.replay_race', race_id=race.id) }}{{ replay_anchor }}">Replay</a>
                 {% endif %}
-                <a class="btn btn-outline-secondary btn-admin btn-sm btn-download" href="{{ url_for('race.download_race_json', race_id=race.id) }}">Download</a>
+                <a class="btn btn-outline-secondary btn-admin btn-sm btn-download" href="{{ url_for('admin_race.download_race_json', race_id=race.id) }}">Download</a>
               </td>
             {% endif %}
             <td class="bool-{{ race.is_visible | lower }}">
@@ -86,7 +86,7 @@
               </td>
             {% else %}
               <td colspan="3" class="text-right">
-                <a href="{{ url_for('race.upload_results', race_id=race.id) }}" class="btn btn-outline-secondary btn-admin btn-sm btn-jump">Upload results</a>
+                <a href="{{ url_for('admin_race.upload_results', race_id=race.id) }}" class="btn btn-outline-secondary btn-admin btn-sm btn-jump">Upload results</a>
               </td>
             {% endif %}
           </tr>
@@ -98,8 +98,8 @@
     <div class="row">
       <div class="col my-3">
         <p class="btn-collection">
-          <a href="{{ url_for('race.new_race')}}" class="btn btn-outline-secondary btn-admin btn-jump">Add new race</a>
-          <a href="{{ url_for('race.show_tracks')}}" class="btn btn-outline-secondary btn-admin btn-jump">View racetracks</a>
+          <a href="{{ url_for('admin_race.new_race')}}" class="btn btn-outline-secondary btn-admin btn-jump">Add new race</a>
+          <a href="{{ url_for('admin_race.show_tracks')}}" class="btn btn-outline-secondary btn-admin btn-jump">View racetracks</a>
         </p>
       </div>
     </div>
