+++ buggy_race_server/templates/admin/racetracks.html
@@ -12,7 +12,7 @@
   </div>
   <div class="row">
     <div class="col btn-collection">
-      <a href="{{ url_for('race.new_track') }}" class="btn btn-admin btn-outline-secondary btn-jump">Add new racetrack</a>
+      <a href="{{ url_for('admin_race.new_track') }}" class="btn btn-admin btn-outline-secondary btn-jump">Add new racetrack</a>
     </div>
   </div>
   {% if is_showing_example_racetracks %}
@@ -22,7 +22,7 @@
         don't see them here, you can automatically add them.
       </div>
       <div class="col-md-6 p-3">
-        <form action="{{ url_for('race.autofill_tracks') }}" method="POST">
+        <form action="{{ url_for('admin_race.autofill_tracks') }}" method="POST">
           {{ form.csrf_token }}
           <button class="btn btn-primary">Add missing racetracks</button>
         </form>
